fix(plugin-changesets): prompt for workspaces included in changesets
**Problem:** When selecting workspaces for versioning, other packages were being versioned based on only the changesets related to the selected workspaces.
This isn't quite right when there's a single changeset that spans multiple workspaces where one or more of those workspaces is not part of the list of workspaces originally requested to be versioned. This results in incorrectly versioning those extra workspaces and not including all of their changesets in changelogs.

**Solution:** Check for the extra workspaces, prompt that it will be necessary to included them, then do so if okay to proceed. Exit early and fail if not okay.