fix(test-cli): remove unnecessary parse/unparse steps in test builder (#473)
**Problem:**

When attempting to run a test with something like `run('-w "foo bar"')`,
the argument parser in tests was resulting in arguments `{ _: ['bar'],
'w': 'foo' }`, but we expect it to be `{ _: [], 'w': 'foo bar' }`

**Solution:**

There was a `parse` + `unparse` step within the `builder` for the
test-cli, likely existed for some historical reasons that are no longer
relevant. This removes them.

Also sneaking in a fix for missing semver dependency.