refactor(core): Workspace commands + passthrough (#591)
**Problem:**

Looking at #554 and #555, there are too many unknowns with passing
through to commands like vite and astro (and webpack, next, etc, for
that matter). It doesn't seem sustainable to attempt to curate plugins
to every single npx command available. Nor does it seem helpful for
users of oneRepo to leave them to need to write the entire passthrough
on their own.

**Solution:**

Passthrough commands. Discussion in #590

**Related issues:**

Fixes #554
Fixes #555

**Checklist:**

- [x] Added or updated tests
- [x] Added or updated documentation
- [x] Ensured the pre-commit hooks ran successfully