feat(core): auto install node_modules if necessary (#579)
**Problem:**

One thing that comes up a bunch is needing to remember to run `npm
install` periodically. If it's not done ahead of trying to run commands
or work with code in the repo, developers can potentially receiving
confusing errors, particularly coming from areas they haven't been
working in.

**Solution:**

At the `one` level before setup, before even requiring the local
`onerepo` package or any of its dependencies:

1. Ensure we are in a valid repo
2. Get the configuration and root of the repository
3. if in `CI` or `ONEREPO_USE_HOOKS="0"`, go to 9
4. Get the lockfile, if not exists, go to 9
5. get a hash of the lockfile's contents
6. check if the lockfile's content hash matches a cache from the last
time we ran this (if it hasn't run, it does not match) if matches, go to
9
7. run `graph.packageManager.install()`
8. write the lockfile's content has to a cache file
9. continue importing `setup` from `'onerepo'` and run commands

Bonus: after splitting up and running the "global" version of `one` if a
valid config/root does not exist, we can throw a better error message to
help developers out when there are syntax errors in commands or other
issues.