fix(eslint): rename quiet option to warnings (#703)
**Problem:**

The `--quiet` flag on `one eslint` is conflicting with the global
`--quiet` flag on the `one` CLI, which sets the logger into silent mode
(no output). This will make eslint look like nothing is happening and
potentially fail silently.

**Solution:**

Switch from `quiet` to `warnings`. This should increase clarity of the
flag and avoid conflicts.

**This is a breaking change!**

**Related issues:**

<!--
- Link the issue or issues being fixed so they are appropriately tracked
and marked closed.
-->

Fixes #

**Checklist:**

- [x] Added or updated tests
- [x] Added or updated documentation
- [x] Ensured the pre-commit hooks ran successfully

_By opening this pull request, you agree that this submission can be
released under the same
[License](https://github.com/paularmstrong/onerepo/blob/main/LICENSE.md)
that covers the project._