feat!: drop node 18 (#814)
**Problem:**

Node 18 is now end of life and not receiving maintenance or security
updates. https://nodejs.org/en/about/previous-releases

**Solution:**

Drop Node 18 support.

New support is based on the [Yargs 18 release
support](https://github.com/yargs/yargs/releases/tag/v18.0.0), as it is
a core dependency and we will look to upgrade to this soon.