feat(onerepo): upgrade yargs to v18 (#815)
**Problem:**

Yargs 18 is available as an ESM-first package. This is a great stepping
stone for us as we look toward the future of removing Jiti as a
dependency for loading TypeScript files starting with Node 22's
`--experimental-strip-types` support.

**Solution:**

Upgrade Yargs. This appears to work as a drop-in replacement with no
real changes required on our end.